Solution day 5 part 1
This question required the option to use immediate and parameter modes
when reading values from memory as well as support negative values. In order
to provide this functionality, the main memory data type needed to be changed
from usize to i64.

+++ src/year_2019/day_2.rs
@@ -4,12 +4,12 @@ use std::fs;
 #[allow(dead_code)]
 pub fn part_1() {
     let location = "input/year_2019/day_2_1.txt";
-    let input: Vec<usize> = fs::read_to_string(location)
+    let input: Vec<i64> = fs::read_to_string(location)
         .expect("Cannot read input file")
         .split(",")
         .map(|s| s.to_string())
-        .filter_map(|s| s.parse::<usize>().ok())
-        .collect::<Vec<usize>>();
+        .filter_map(|s| s.parse::<i64>().ok())
+        .collect::<Vec<i64>>();
     let mut intcode = Intcode::new(input).expect("Input cannot be intialized");
     intcode.reset();
     intcode.write_noun(12);
@@ -23,12 +23,12 @@ pub fn part_1() {
 #[allow(dead_code)]
 pub fn part_2() {
     let location = "input/year_2019/day_2_1.txt";
-    let input: Vec<usize> = fs::read_to_string(location)
+    let input: Vec<i64> = fs::read_to_string(location)
         .expect("Cannot read input file")
         .split(",")
         .map(|s| s.to_string())
-        .filter_map(|s| s.parse::<usize>().ok())
-        .collect::<Vec<usize>>();
+        .filter_map(|s| s.parse::<i64>().ok())
+        .collect::<Vec<i64>>();
     let mut intcode = Intcode::new(input).expect("Input cannot be intialized");
     for noun in 0..=99 {
         for verb in 0..=99 {
